requirements.txt
PyQt6
opencv-python
ffmpeg-python
numpy
google-generativeai
Pillow
send2trash
pyspellchecker==0.8.1